// types/Team.ts
export interface Team {
	code: number;
	draw: number;
	form: number | null;
	id: number;
	loss: number;
	name: string;
	played: number;
	points: number;
	position: number;
	short_name: string;
	strength: number;
	team_division: number | null;
	unavailable: boolean;
	win: number;
	strength_overall_home: number;
	strength_overall_away: number;
	strength_attack_home: number;
	strength_attack_away: number;
	strength_defence_home: number;
	strength_defence_away: number;
	pulse_id: number;
}